% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myreg.R
\name{myreg}
\alias{myreg}
\title{Perform linear regression using the normal equation learnt in BIOS 650 class.}
\usage{
myreg(formula, data)
}
\arguments{
\item{formula}{A formula specifying the outcome and predictor variables.}

\item{data}{A data frame containing all the variables in the formula.}
}
\description{
Perform linear regression using the normal equation learnt in BIOS 650 class.
}
