% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myreg.R
\name{summary.myreg}
\alias{summary.myreg}
\title{Summarize a linear regression model}
\usage{
\method{summary}{myreg}(model)
}
\arguments{
\item{model}{An object of class "myreg".}
}
\description{
Summarize a linear regression model
}
