% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myreg.R
\name{summary.myreg}
\alias{summary.myreg}
\title{Summarize a linear regression model}
\usage{
\method{summary}{myreg}(model)
}
\arguments{
\item{model}{An object of class "myreg".}
}
\description{
Summary.myreg function provides a summary of the results from a linear regression model fitted using myreg function.
The summary includes estimated coefficients, their standard errors, t-values, p-values,
R-squared, adjusted R-squared, and the F-statistic with its p-value.
}
\examples{
# Fit a regression model
iris_model <- myreg(Sepal.Length ~ Sepal.Width + Species, data = iris)

# Summarize the regression model
summary(iris_model)
}
